Attempted to implement saving Location data tot the history in SavedLocations.java Attempts are commented out Changed the way the values of theHistory table in the rows and columns

+++ app/src/main/java/com/example/ronald/sfparking/Park_LocationDataSource.java
@@ -36,9 +36,9 @@ public class Park_LocationDataSource {
             cursor.moveToNext();
 
         }
-        public LocationInfo getLocationInfo(long id){
+        public LocationInfo getLocationInfo(int id){
             Cursor cursor = database.query(MySQLiteHelper.TABLE_PARKINFO, allColumns, MySQLiteHelper.KEY_ID + " = " + new String []{String.valueOf(id)}, null, null, null, null );
-            cursor.moveToPosition((int)id);
+            cursor.moveToPosition(id);
             LocationInfo locationInfo = new LocationInfo(cursor.getInt(0), cursor.getDouble(1), cursor.getDouble(2), cursor.getString(3), cursor.getString(4),cursor.getString(5));
             return locationInfo;
         }
